Move H5Completer class back into h5glance.terminal module
